requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
asyncio==3.4.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comtypes==1.4.1
disutils==1.4.32.post2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
MouseInfo==0.1.3
multidict==6.0.5
pillow==10.3.0
pip==24.0
psutil==5.9.8
PyAudio==0.2.14
PyAutoGUI==0.9.54
pygame==2.5.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
pyttsx3==2.90
pytweening==1.2.0
pywin32==306
requests==2.31.0
setuptools==69.5.1
sniffio==1.3.1
SpeechRecognition==3.10.3
tk==0.1.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
youtube-search-python==1.6.6